+++ src/utils.cc
@@ -114,7 +114,7 @@ timespec get_uptime() {
         now.tv_sec -= time_sync_data.hi_res_clock.tv_sec;
         now.tv_nsec -= time_sync_data.hi_res_clock.tv_nsec;
         if (now.tv_nsec < 0) {
-            now.tv_nsec += 1e9;
+            now.tv_nsec += (long long int)1e9;
             now.tv_sec--;
         }
         return now;
@@ -132,7 +132,7 @@ precise_time_t get_absolute_time(const timespec& relative_time) {
     time_t sec = time_sync_data.low_res_clock + relative_time.tv_sec;
     uint32_t nsec = time_sync_data.hi_res_clock.tv_nsec + relative_time.tv_nsec;
     if (nsec > 1e9) {
-        nsec -= 1e9;
+        nsec -= (long long int)1e9;
         sec++;
     }
     (void) gmtime_r(&sec, &result);
